// src/app/task-detail/modules/sidebar-right/modules/widget-rent-manager/modules/rent-manager-notes/enums/rent-manager-notes.enum.ts
export enum ERentManagerNotesPopup {
  SELECT_NOTES = 'SELECT_NOTES',
  SYNC_NOTES = 'SYNC_NOTES',
  ATTACH_FILE = 'ATTACH_FILE'
}
export enum ERentManagerNotesType {
  CREATE_NEW = 'CREATE_NEW',
  SELECT_EXISTING = 'SELECT_EXISTING'
}

export enum ENoteSaveToType {
  TENANT = 'Tenancy',
  OWNERSHIP = 'Ownership',
  PROPERTY = 'Property'
}
export enum ECategoryType {
  ALL = 'ALL',
  TENANT = 'TENANT',
  OWNER = 'OWNER',
  PROPERTY = 'PROPERTY',
  UNIT = 'UNIT'
}

export enum EEntityType {
  RM_CRM_CURRENT = 'RM_CRM_CURRENT',
  RM_CRM_PAST = 'RM_CRM_PAST',
  DELETED = 'DELETED',
  PROSPECT = 'PROSPECT'
}